Restructure commands onto their own @Command AutoCollect pipeline
