+++ src/main/java/com/github/mkram17/bazaarutils/commands/HelpCommand.java
@@ -1,10 +1,8 @@
 package com.github.mkram17.bazaarutils.commands;
 
 import com.github.mkram17.bazaarutils.generated.BazaarUtilsLateInitModules;
-import com.github.mkram17.bazaarutils.generated.BazaarUtilsModules;
 import com.github.mkram17.bazaarutils.utils.PlayerActionUtil;
-import com.github.mkram17.bazaarutils.utils.Util;
-import com.github.mkram17.bazaarutils.utils.annotations.modules.Module;
+import com.github.mkram17.bazaarutils.utils.annotations.modules.Command;
 import com.mojang.brigadier.builder.LiteralArgumentBuilder;
 import com.mojang.brigadier.context.CommandContext;
 import lombok.Getter;
@@ -13,9 +11,7 @@ import net.minecraft.network.chat.MutableComponent;
 import net.minecraft.network.chat.Component;
 import net.minecraft.ChatFormatting;
 
-import java.util.List;
-
-@Module
+@Command
 public final class HelpCommand implements BUCommand {
 
     @Getter public final String commandName = "help";
@@ -27,7 +23,7 @@ public final class HelpCommand implements BUCommand {
     }
 
     private int execute(CommandContext<FabricClientCommandSource> context) {
-        BazaarUtilsCommands root = BazaarUtilsLateInitModules.BazaarUtilsCommands;
+        BazaarUtilsCommand root = BazaarUtilsLateInitModules.BazaarUtilsCommand;
 
         MutableComponent message = Component.literal("BazaarUtils Commands\n").withStyle(ChatFormatting.GOLD);
