+++ src/main/java/com/github/mkram17/bazaarutils/commands/orders/RemoveOrderCommand.java
@@ -1,0 +1,41 @@
+package com.github.mkram17.bazaarutils.commands.orders;
+
+import com.github.mkram17.bazaarutils.commands.BUCommand;
+import com.github.mkram17.bazaarutils.commands.OrdersCommands;
+import com.github.mkram17.bazaarutils.misc.NotificationType;
+import com.github.mkram17.bazaarutils.utils.PlayerActionUtil;
+import com.github.mkram17.bazaarutils.utils.annotations.modules.Command;
+import com.github.mkram17.bazaarutils.utils.bazaar.market.order.Order;
+import com.github.mkram17.bazaarutils.utils.storage.UserOrdersStorage;
+import com.mojang.brigadier.arguments.IntegerArgumentType;
+import com.mojang.brigadier.builder.LiteralArgumentBuilder;
+import com.mojang.brigadier.context.CommandContext;
+import lombok.Getter;
+import net.fabricmc.fabric.api.client.command.v2.ClientCommandManager;
+import net.fabricmc.fabric.api.client.command.v2.FabricClientCommandSource;
+import net.minecraft.ChatFormatting;
+import net.minecraft.network.chat.Component;
+
+@Command(parent = OrdersCommands.class)
+public final class RemoveOrderCommand implements BUCommand {
+    @Getter
+    public final String commandName = "remove";
+
+    @Getter
+    public final Component description = Component.literal("Removes an order by index.").withStyle(ChatFormatting.GRAY);
+
+    @Override
+    public LiteralArgumentBuilder<FabricClientCommandSource> getCommandBuilder(LiteralArgumentBuilder<FabricClientCommandSource> base) {
+        return base.then(ClientCommandManager.argument("index", IntegerArgumentType.integer()).executes(this::removeByIndex));
+    }
+
+    private int removeByIndex(CommandContext<FabricClientCommandSource> context) {
+        int index = IntegerArgumentType.getInteger(context, "index");
+
+        Order order = UserOrdersStorage.INSTANCE.get().get(index);
+        order.removeFromUserOrders();
+        PlayerActionUtil.notifyAll("Removed " + order, NotificationType.COMMAND);
+
+        return 1;
+    }
+}
